requirements.txt
# Core
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web Scraping
firecrawl-py>=1.0.0

# Backend API
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Voice
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0

# Evaluation
ragas>=0.1.10
datasets>=2.20.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Utilities
httpx>=0.27.0
aiofiles>=24.1.0
tqdm>=4.66.0
